feat: 부모 프레임 클릭 시에도 모달이 닫히는 기능 추가
