+++ frontend/deploy-script/src/index.js
@@ -1,22 +1,26 @@
+import { createIframe } from "./createIframe.js";
+
 (function () {
   const $darass = document.querySelector("#darass");
 
   const replyModuleURL = "https://dorvcm7xtbd6v.cloudfront.net" + "?";
-  const currURL = (window.location.origin + window.location.pathname).replace("index.html", "");
+  const currentURL = (window.location.origin + window.location.pathname).replace("index.html", "");
 
   const projectKey = $darass.dataset.projectKey;
 
   const urlParams = new URLSearchParams(replyModuleURL);
-  urlParams.set("url", currURL);
+  urlParams.set("url", currentURL);
   urlParams.set("projectKey", projectKey);
 
-  const $iframe = document.createElement("iframe");
-  const defaultStyle = `width: 1px !important; min-width: 100% !important; border: none !important; overflow: hidden !important; height: 540px !important;`;
-  $iframe.setAttribute("scrolling", "no");
-  $iframe.setAttribute("src", decodeURIComponent(urlParams.toString()));
-  $iframe.setAttribute("style", defaultStyle);
+  const url = decodeURIComponent(urlParams.toString());
+  const $iframe = createIframe(url);
   $darass.appendChild($iframe);
 
+  const postMessageToIframe = () => {
+    $iframe.contentWindow.postMessage("click", "*");
+  };
+
+  window.addEventListener("click", postMessageToIframe);
   window.addEventListener("message", event => {
     if (typeof event.data === "number") {
       $iframe.style.setProperty("height", `${event.data}px`, "important");
