+++ frontend/reply-module/src/components/atoms/Modal/index.tsx
@@ -1,4 +1,4 @@
-import { ReactNode } from "react";
+import { ReactNode, useEffect } from "react";
 import ReactDOM from "react-dom";
 import { Container, Dimmed } from "./styles";
 
@@ -10,6 +10,14 @@ export interface Props {
 const $modalRoot = document.getElementById("modal-root");
 
 const Modal = ({ children, onCloseModal }: Props) => {
+  useEffect(() => {
+    window.addEventListener("click", onCloseModal);
+
+    return () => {
+      window.removeEventListener("click", onCloseModal);
+    };
+  }, []);
+
   if (!$modalRoot) return null;
 
   return ReactDOM.createPortal(
